requirements.txt
Django==4.1.4
psycopg2-binary==2.9.5
django-environ==0.9.0 
djangorestframework==3.14.0
drf-spectacular==0.25.1 